Split quick phrases UI into focused component files
Move sub-components out of the monolithic quick-phrases-button.tsx into a
quick-phrases/ subfolder: edit-phrase-dialog, fill-variables-dialog,
phrase-row, inject-split-button. The main quick-phrases-button.tsx now
contains only orchestration (~180 lines). Re-exported via index.ts.
