+++ apps/web/src/components/app/quick-phrases/index.ts
@@ -1,0 +1,1 @@
+export { QuickPhrasesButton } from "./quick-phrases-button";
